requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch 
torchvision 
numpy
ipython 
ipykernel
pyyaml
plotly
nbformat>=4.2.0
tqdm
mediapipe